Resolve user guide CSS in 1 request
This combines 4 files into 1. base.css now contains
docs.css and userguide.css as well as inlines the content for the
link icon.

This will allow other requests to load in parallel, thus reducing
the overall page load time significantly.

Display of all pages should be the same, as styles were changed to
be mutually exclusive.

Signed-off-by: Eric Wendelin <eric@gradle.com>